Express frequencies in Hz instead of MHz

+++ include/Rfm69/Rfm69.h
@@ -9,7 +9,7 @@ namespace Rfm69
 {
 
 using NodeAddress = uint8_t;
-using Frequency   = float;
+using Frequency   = uint32_t;
 using BitRate     = uint32_t;
 
 struct Frame
